Finished refactoring Scrapes page table

+++ helpers/parseCSV.js
@@ -25,7 +25,7 @@ class parseCSV {
             if (lineIndex == 0) {
               return;
             }
-            obj.values.push(parseInt(line[clientIndex]));
+            obj.values.push(parseFloat(line[clientIndex]));
           });
 
           acc.push(obj);
